Error incorrect hash while installing with ES 2.0
We are moving back release scripts locally as it's the last repository which is not in elasticsearch repo.
Also, we now use generated checksums when we run the `install` phase and we upload them to S3 at the same time we upload the artifact itself.

Closes #178.

(cherry picked from commit d3324bc)
(cherry picked from commit 311a80b)